backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.9.2
langchain-core==0.3.15
langchain-community==0.3.4
langchain-google-genai==2.0.5
python-dotenv==1.0.1
cryptography==43.0.3
